requirements.txt
django-json-widget==1.0.0
jsonformatter==0.1.3
